Change Int64 to Int

+++ src/models/OuterControl.jl
@@ -7,7 +7,7 @@
         reactive_power::R
         ext::Dict{String, Any}
         states::Vector{Symbol}
-        n_states::Int64
+        n_states::Int
     end
 Parameters of a Outer-Loop controller using a active power controller and a reactive power droop controller.
 
@@ -16,7 +16,7 @@ Parameters of a Outer-Loop controller using a active power controller and a reac
 - `R <: ReactivePowerControl`: Reactive power controller (typically droop).
 - `ext::Dict{String, Any}`
 - `states::Vector{Symbol}`: Vector of states (will depend on the components).
-- `n_states::Int64`: Number of states (will depend on the components).
+- `n_states::Int`: Number of states (will depend on the components).
 """
 mutable struct OuterControl{A <: ActivePowerControl, R <: ReactivePowerControl} <:
                DynamicInverterComponent
@@ -24,7 +24,7 @@ mutable struct OuterControl{A <: ActivePowerControl, R <: ReactivePowerControl} 
     reactive_power::R
     ext::Dict{String, Any}
     states::Vector{Symbol}
-    n_states::Int64
+    n_states::Int
 end
 
 function OuterControl(
